Add an options menu to the title screen
This allows the player to enable/disable animations, music, and sound
effects. These settings are then stored in a cookie (seperate from the
one used to store the player's game state).

Menus are navigated with up/down and the action button.
Alternatively, it's possible to click directly on menu entries to
activate them.